Add workflow for edge build of RFID LLRP device and app snaps
Add minimal workflow files for

- wf-device-rfid-llrp
- wf-rfid-llrp-inventory

These workflows do a daily build of the snaps and if they succeed, then kick off
a build on Launchpad, which will in turn publish a snap to latest/edge

As the snaps are still in edgexfoundry-holding, syncing the repositories is
disabled for now. As the upstream snaps don't have a version tag,
the snaps will also be version 0.0.0

When the upstream repositories get moved into the edgexfoundry organization then
these workflows will be updated to process the sync and versioning correctly.

Signed-off-by: Siggi Skulason <siggi.skulason@canonical.com>